dcpu-ide: Begin proper implementation of modal dialog contents.
This includes the ability to (re)set the dialog title and
body contents as well as supply definitions for all the
buttons we should include in the dialog. These buttons are
pre-determined by constants. This makes it easier to
organize their placement in the dialog as we already know
what each button will be used for.