dcpu-ide: Fixes implementation of Dialog class.
This adds a couple more button types we can use.
Additionally added some comments, noting that this is the most
basic type of Dialog we have and should probably never be
used directly. The intention is for subclasses to specialize
the behaviour of a Dialog for specific purposes.

Examples would be a Confirmation dialog, information display,
error dialog or a wizard of some kind.