dcpu-ide: Major overhaul of Makefile and codegen tool.
The codegen tool is now invoked for each code generation instance
separately. This keeps the tool simpler and cleaner and makes
the Makefile more descriptive as to what exactly is being done by
the tool. Moved codegen templates to the templates subdirectory.
Added index.html to code generation stage to have its list of
included stylesheets and scripts automatically generated.

This will allow us to merge all of these into a single .css and .js
file for release builds. This will greatly reduce the amount of
roundtrips the web app needs to make upon load.