dcpu-ide: Fixes api.request to return an ApiError object on errors. Fixes implementation of createProject.
When api.request is run in async mode, the existing error handler
is still applied. When run in sync mode, it will throw an ApiError
exception which can be caught in a try/catch block. This simplifies
the invocation code for api.request somewhat.

createProject now checks if an existing project is already open
and whether or not it has unsaved changes. A notification is given
if this is the case and gives the user the choice to continue or
stop the project creation.