dcpu-ide: Fix codegen and Makefile's merge stage.
This fixes the merge stage for the install build. It now adheres to
the originally defined order of javascript and css files as they
are written to the unified files. Instead of traversing all css/js
files and merging them as they come along, it loads codege.json
and uses the file lists from there. It also deletes these files once
they have been merged.