dcpu-ide: Implement the rest of apiNewProject. Implement apiDirList.
apiNewProject creates the new project directory and copies over
templates source files which form the basis for a new project.
This includes a single `main.dasm` and a `README.md`. Template
variables inside these are replaced with concrete values from
the configuration. These can refer to any field in the Project
struct.

apiDirList returns a flat list of file/directory paths. Each with
a boolean flag indication whether it is a directory or not.
Added an optional recursion flag to indicate if should traverse
subdirectories or not. It limits the returned data to anything
inside config.ProjectPath (The base directory for all code projects).